feat: [PAGOPA-3061] Renamed package
