+++ src/main/java/it/gov/pagopa/gpd/rtp/model/AppCorsConfiguration.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.model;
+package it.gov.pagopa.gpd.rtp.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
