+++ src/main/java/it/gov/pagopa/gpd/rtp/util/Constants.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.util;
+package it.gov.pagopa.gpd.rtp.util;
 
 import lombok.experimental.UtilityClass;
