+++ src/test/java/it/gov/pagopa/gpd/rtp/ApplicationTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice;
+package it.gov.pagopa.gpd.rtp;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
